Remove redundant RuboCop config
Previously we disabled the HelperInstanceVariable Cop for the entire
repo on the basis of a RedirectHelper class that's actually a service
in disguise - it's certainly not used in a view. This moves the helper
into the services/ directory and renames it as such.
